Server logic endpoints

+++ server/db/migrations/20221101081143_users.js
@@ -15,6 +15,12 @@ exports.up = function(knex) {
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = function(knex) {
-    return knex.schema.dropTable('users');
+exports.down = async function(knex) {
+    await knex.schema
+    .hasTable('users')
+    .then(async function(exists) {
+        if (exists) {
+            return knex.schema.dropTable('users');
+        };
+    });
 };
